Refactor redis lock to extract common code to be used also in redis queue.
